+++ resources/views/errors/Error.fortune.php
@@ -3,6 +3,6 @@
 <% part("content") %>
     <h2><% show("errorTitle") %></h2>
     <% part("errorDescription") %>
-        Something went wrong.
+        Something went wrong.  We will look into what happened.
     <% show %>
 <% endpart %>